refactor_composite-primitves3d
